This passes tests.

I haven't yet fed a full xul.sym to into it. This is probably a good opportunity to try out socc-pair.