@Gankra 2022-06-29
**None of the following is needed or requested**, errors for this parser aren't even user-facing. I've just been in a deep hole this month trying to figure out how to get good parsing errors, and wanted to spread the good word in case it's ever useful.

One little tip that I didn't understand until recently: when you have an `alt` your errors turn to garbage because nom can't distinguish between "didn't match this pattern" and "did match the pattern but the subpattern was corrupt". You can use [cut](https://docs.rs/nom/latest/nom/combinator/fn.cut.html) to specify "at this point, I am confident I have matched the pattern that an outer alt is checking for, so if the parse fails then the input is corrupt and this is to blame".

So basically if you toss a `cut` after each line's initial `tag` you can properly detect which line variant failed to parse. You can further use [the context combinator and convert_error](https://github.com/Geal/nom/blob/main/doc/error_management.md) to add better descriptions of what went wrong in the parse.