@luser 2022-06-29
Nice work! I don't know if the tests here are complete enough to rely on, but we do have fuzz targets in this crate as well, so for additional confidence you could try `cargo fuzz`ing it.