@mstange 2022-06-29
I fixed the `m` space issue and added some calls to `cut` but I didn't add any nice error messages.

And I fed [this xul.sym](https://symbols.mozilla.org/xul.pdb/2E7FA147D9CC4B434C4C44205044422E1/xul.sym) into it and it worked.

I also found out that using socc-pair is a lot less fun if you don't have crash-stats minidump access or a bunch of existing minidumps on your machine.