@Gankra 2022-06-29
> I also found out that using socc-pair is a lot less fun if you don't have crash-stats minidump access or a bunch of existing minidumps on your machine.

Oh whoops, I thought that permission was super normal to have. Is it not?