@mstange 2022-06-29
Maybe it is, I'm not sure. I don't spend a lot of time debugging crashes so I've never asked for access. Anyway, Jeff downloaded a dump for me and I'm currently hacking up socc-pair so that I can compare a local minidump with a socorro JSON.