@gabrielesvelto 2022-06-30
> Oh whoops, I thought that permission was super normal to have. Is it not?

It's rather uncommon because of the toxic nature of the data. @mstange do request protected data acccess, it will make your life a lot easier and given you've been doing a lot of work in this area it's definitely something you need. The instructions are [here](https://crash-stats.mozilla.org/documentation/protected_data_access/).