@Gankra 2022-06-30
Benchmark results:

# Summary

runtime:
* **cached: 1762ms -> 2084ms (oof!)**
* http: 6482ms -> 6590ms

memory usage unchanged

# methodology

```
> socc-pair --bench=5  --run-local=<snip> --api-token=<snip> --crash-id=d0170905-001e-430d-a664-452cc0220630 --mock-server --output-file=results-main-serverless.txt
> socc-pair --bench=5  --run-local=<snip> --api-token=<snip> --crash-id=d0170905-001e-430d-a664-452cc0220630 --clean-cache --output-file=results-main-download.txt

> git checkout nom7  

> socc-pair --bench=5  --run-local=<snip> --api-token=<snip> --crash-id=d0170905-001e-430d-a664-452cc0220630 --clean-cache --output-file=results-nom7-download.txt
> socc-pair --bench=5 --run-local=<snip> --api-token=<snip> --crash-id=d0170905-001e-430d-a664-452cc0220630 --mock-server --output-file=results-nom7-serverless.txt
```


# main, serverless (--mock-server)

```
benchmark results (ms):
  1750, 1773, 1798, 1758, 1735, 
average runtime: 00m:01s:762ms (1762ms)
median runtime: 00m:01s:758ms (1758ms)
min runtime: 00m:01s:735ms (1735ms)
max runtime: 00m:01s:798ms (1798ms)

max memory (rss) results (bytes):
  260980736, 260538368, 261054464, 260374528, 260497408, 
average max-memory: 248MB (260689100 bytes)
median max-memory: 248MB (260538368 bytes)
min max-memory: 248MB (260374528 bytes)
max max-memory: 248MB (261054464 bytes)
```

# nom7, serverless (--mock-server)

```
benchmark results (ms):
  2146, 2093, 2077, 2087, 2021, 
average runtime: 00m:02s:084ms (2084ms)
median runtime: 00m:02s:087ms (2087ms)
min runtime: 00m:02s:021ms (2021ms)
max runtime: 00m:02s:146ms (2146ms)

max memory (rss) results (bytes):
  259751936, 260878336, 260886528, 260718592, 260202496, 
average max-memory: 248MB (260487577 bytes)
median max-memory: 248MB (260718592 bytes)
min max-memory: 247MB (259751936 bytes)
max max-memory: 248MB (260886528 bytes)
```

# main, download (--clean-cache)

```
benchmark results (ms):
  6855, 6562, 6264, 6357, 6375, 
average runtime: 00m:06s:482ms (6482ms)
median runtime: 00m:06s:375ms (6375ms)
min runtime: 00m:06s:264ms (6264ms)
max runtime: 00m:06s:855ms (6855ms)

max memory (rss) results (bytes):
  261070848, 259944448, 260919296, 260550656, 260378624, 
average max-memory: 248MB (260572774 bytes)
median max-memory: 248MB (260550656 bytes)
min max-memory: 247MB (259944448 bytes)
max max-memory: 248MB (261070848 bytes)
```

# nom7, download (--clean-cache)

```
benchmark results (ms):
  6504, 6724, 6665, 6503, 6555, 
average runtime: 00m:06s:590ms (6590ms)
median runtime: 00m:06s:555ms (6555ms)
min runtime: 00m:06s:503ms (6503ms)
max runtime: 00m:06s:724ms (6724ms)

max memory (rss) results (bytes):
  261120000, 261070848, 261193728, 261021696, 260825088, 
average max-memory: 248MB (261046272 bytes)
median max-memory: 248MB (261070848 bytes)
min max-memory: 248MB (260825088 bytes)
max max-memory: 249MB (261193728 bytes)
```