@Gankra 2022-06-30
I expected that benchmark to be a formality, but that's quite a regression. It's something we *could* swallow but I don't feel great about it given this isn't adding functionality or fixing some known bugs (?)