@Gankra 2022-06-30
Hmm ok actually I actually computed the multiplier and it's "just" 1.18x. Raw numbers seemed more dramatic. Still not *great* but maybe something we could eat since rust-minidump already halved these numbers over breakpad.