@mstange 2022-06-30
I've added some commits to optimize things. Tests still pass.

On my machine, the commits reduce the time from 991ms to 737ms, in the following steps:

991ms -> 901ms -> 856ms -> 752ms -> 737ms